pyproject.toml
[tool.poetry]
name = "news-interview"
version = "0.1.0"
description = ""
authors = ["AdmiralDesu <skyrim.men024@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiosqlite = "^0.17.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.37"}
sqlmodel = "^0.0.6"
Flask-RESTful = "^0.3.9"
flask-apispec = "0.11.0"
bs4 = "^0.0.1"
Flask-SQLAlchemy = "^2.5.1"
marshmallow-sqlalchemy = "^0.28.0"
flask-marshmallow = "^0.14.0"
requests = "^2.28.0"
lxml = "^4.9.0"
rq-scheduler = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
